refactor: Reservation 관리자, 사용자 컨트롤러 분리
