+++ src/main/java/roomescape/reservation/controller/ReservationAdminController.java
@@ -1,0 +1,40 @@
+package roomescape.reservation.controller;
+
+import java.util.List;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.reservation.dto.ReservationResponse;
+import roomescape.reservation.dto.ReservationSaveDto;
+import roomescape.reservation.service.ReservationService;
+
+@RestController
+@RequestMapping("/admin")
+public class ReservationAdminController {
+    private final ReservationService reservationService;
+
+    public ReservationAdminController(ReservationService reservationService) {
+        this.reservationService = reservationService;
+    }
+
+    @GetMapping("/reservations")
+    public ResponseEntity<List<ReservationResponse>> read() {
+        return ResponseEntity.ok(reservationService.findAll());
+    }
+
+    @PostMapping("/reservations")
+    public ResponseEntity<ReservationResponse> create(@RequestBody ReservationSaveDto dto) {
+        return ResponseEntity.ok(reservationService.create(dto));
+    }
+
+    //TODO: status 변경으로 수정
+    @DeleteMapping("/reservations/{id}")
+    public ResponseEntity<ReservationResponse> delete(@PathVariable Long id) {
+        return ResponseEntity.ok(reservationService.delete(id));
+    }
+}
