+++ src/main/java/roomescape/reservation/controller/ReservationController.java
@@ -1,39 +1,12 @@
 package roomescape.reservation.controller;
 
-import java.util.List;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.reservation.dto.ReservationSaveDto;
-import roomescape.reservation.dto.ReservationResponse;
-import roomescape.reservation.service.ReservationService;
 
 @RestController
-@RequestMapping("/reservations")
+@RequestMapping("/member")
 public class ReservationController {
-    private final ReservationService reservationService;
-
-    public ReservationController(ReservationService reservationService) {
-        this.reservationService = reservationService;
-    }
-
-    @GetMapping
-    public ResponseEntity<List<ReservationResponse>> read() {
-        return ResponseEntity.ok(reservationService.findAll());
-    }
-
-    @PostMapping
-    public ResponseEntity<ReservationResponse> create(@RequestBody ReservationSaveDto dto) {
-        return ResponseEntity.ok(reservationService.create(dto));
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<ReservationResponse> delete(@PathVariable Long id) {
-        return ResponseEntity.ok(reservationService.delete(id));
-    }
+    // TODO: POST /member/reservations - 사용자 예약 생성
+    // TODO: GET /member/reservations - 사용자 예약 조회 (username 기준)
+    // TODO: PATCH /member/reservations/{id} - 사용자 예약 상태 변경 (취소)
 }
