+++ src/test/java/roomescape/MissionStepTest.java
@@ -16,7 +16,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.reservation.controller.ReservationController;
+import roomescape.reservation.controller.ReservationAdminController;
 import roomescape.reservation.domain.Reservation;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
@@ -142,13 +142,13 @@ class MissionStepTest {
     }
 
     @Autowired
-    private ReservationController reservationController;
+    private ReservationAdminController reservationAdminController;
 
     @Test
     void 계층화_리팩터링() {
         boolean isJdbcTemplateInjected = false;
 
-        for (Field field : reservationController.getClass().getDeclaredFields()) {
+        for (Field field : reservationAdminController.getClass().getDeclaredFields()) {
             if (field.getType().equals(JdbcTemplate.class)) {
                 isJdbcTemplateInjected = true;
                 break;
